testing : start moving test fixtures to testing package.
