+++ tests/app/preferences/test_prefs.py
@@ -19,8 +19,8 @@ from configobj import ConfigObj
 from future.builtins import str
 from enaml.workbench.api import Workbench
 
-from ...util import (handle_dialog, ecpy_path)
-from ...conftest import APP_DIR_CONFIG, APP_PREFERENCES
+from ecpy.testing.util import (handle_dialog, ecpy_path, APP_DIR_CONFIG,
+                               APP_PREFERENCES)
 
 with enaml.imports():
     from enaml.workbench.core.core_manifest import CoreManifest
