+++ tests/tasks/test_base_views.py
@@ -27,22 +27,22 @@ with enaml.imports():
     from ecpy.tasks.manager.widgets.building import BuilderView
 
 from .manager.conftest import task_workbench
-from ..util import show_widget, process_app_events, handle_dialog, get_window
+from ecpy.testing.util import (show_widget, process_app_events, handle_dialog,
+                               get_window)
 
 
 @pytest.mark.ui
-def test_root_view(windows, task_workbench):
+def test_root_view(windows, task_workbench, dialog_sleep):
     """Test the behavior of the root task view.
 
     """
-    from ..conftest import DIALOG_SLEEP
     task = RootTask()
     view = RootTaskView(task=task,
                         core=task_workbench.get_plugin('enaml.workbench.core'))
     editor = view.children[0]
 
     show_widget(view)
-    sleep(DIALOG_SLEEP)
+    sleep(dialog_sleep)
     assert editor.task is task
     assert editor.root is view
 
@@ -60,18 +60,18 @@ def test_root_view(windows, task_workbench):
     assert task.children
     assert type(task.children[0]) is ComplexTask
     assert len(editor._children_buttons) == 1
-    sleep(DIALOG_SLEEP)
+    sleep(dialog_sleep)
 
     TASK_NAME = 'Bar'
     with handle_dialog('accept', answer_dialog, cls=BuilderView):
         editor.operations['add'](0, 'after')
     process_app_events()
-    sleep(DIALOG_SLEEP)
+    sleep(dialog_sleep)
 
     task.children[0].add_child_task(0, ComplexTask(name='Test'))
     get_window().maximize()
     process_app_events()
-    sleep(DIALOG_SLEEP)
+    sleep(dialog_sleep)
 
     editor.operations['move'](0, 1)
 
@@ -83,11 +83,10 @@ def test_root_view(windows, task_workbench):
 
 
 @pytest.mark.ui
-def test_swapping(windows, task_workbench):
+def test_swapping(windows, task_workbench, dialog_sleep):
     """Test moving a view between containers.
 
     """
-    from ..conftest import DIALOG_SLEEP
     task = RootTask()
     view = RootTaskView(task=task,
                         core=task_workbench.get_plugin('enaml.workbench.core'))
@@ -104,14 +103,14 @@ def test_swapping(windows, task_workbench):
     view.refresh()
     process_app_events()
     assert cont.children == [view]
-    sleep(DIALOG_SLEEP)
+    sleep(dialog_sleep)
 
     view.set_parent(None)
     subview.set_parent(cont)
     subview.refresh()
     process_app_events()
     assert cont.children == [subview]
-    sleep(DIALOG_SLEEP)
+    sleep(dialog_sleep)
 
     subview.set_parent(None)
     view.set_parent(cont)
@@ -119,4 +118,4 @@ def test_swapping(windows, task_workbench):
     process_app_events()
     assert cont.children == [view]
     assert subview.visible
-    sleep(DIALOG_SLEEP)
+    sleep(dialog_sleep)
